Switch main hook from pretxnchangegroup to changegroup
We can't create RB review requrests from pretxnchangegroup,
because RB tries to look up changestes in the repository and
fails as the changesets being pushed are not yet visible from
outside of the pretxnchangegroup hook process.

So we use pretxnchangegroup just to validate the presence of
BUG env var and abort early.